+++ cmd/smailnail/main.go
@@ -11,6 +11,7 @@ import (
 	"github.com/go-go-golems/glazed/pkg/help"
 	help_cmd "github.com/go-go-golems/glazed/pkg/help/cmd"
 	"github.com/go-go-golems/smailnail/cmd/smailnail/commands"
+	annotatecommands "github.com/go-go-golems/smailnail/cmd/smailnail/commands/annotate"
 	enrichcommands "github.com/go-go-golems/smailnail/cmd/smailnail/commands/enrich"
 	smailnaildocs "github.com/go-go-golems/smailnail/cmd/smailnail/docs"
 	"github.com/spf13/cobra"
@@ -95,6 +96,13 @@ func main() {
 		os.Exit(1)
 	}
 	rootCmd.AddCommand(enrichCmd)
+
+	annotateCmd, err := annotatecommands.NewAnnotateCommand()
+	if err != nil {
+		fmt.Printf("Error creating annotate command group: %v\n", err)
+		os.Exit(1)
+	}
+	rootCmd.AddCommand(annotateCmd)
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()